requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.7.1
    # via langchain
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   elastic-transport
    #   requests
    #   sentry-sdk
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via pip-tools
dataclasses-json==0.6.1
    # via langchain
django==4.2.3
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-ninja
django-extensions==3.2.3
    # via -r requirements.in
django-ninja==0.22.2
    # via -r requirements.in
elastic-transport==8.4.1
    # via elasticsearch
elasticsearch==8.10.0
    # via elasticsearch-dsl
elasticsearch-dsl==8.9.0
    # via -r requirements.in
exceptiongroup==1.1.3
    # via anyio
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.305
    # via -r requirements.in
langsmith==0.0.41
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.7
    # via langchain
numpy==1.26.0
    # via
    #   langchain
    #   numexpr
openai==0.28.0
    # via -r requirements.in
orjson==3.9.7
    # via -r requirements.in
packaging==23.1
    # via
    #   build
    #   marshmallow
pip-tools==7.3.0
    # via -r requirements.in
psycopg2-binary==2.9.7
    # via -r requirements.in
pydantic==1.10.12
    # via
    #   django-ninja
    #   langchain
    #   langsmith
pyjwt==2.8.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via elasticsearch-dsl
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via langchain
regex==2023.8.8
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langsmith
    #   openai
    #   tiktoken
sentry-sdk==1.31.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.21
    # via langchain
sqlparse==0.4.4
    # via django
tenacity==8.2.3
    # via langchain
tiktoken==0.5.1
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via
    #   asgiref
    #   pydantic
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==1.26.16
    # via
    #   elastic-transport
    #   requests
    #   sentry-sdk
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
